Many changes.

+++ ax_hmap.c
@@ -20,11 +20,10 @@ struct map_node_t {
 	map_node_t* next;
 };
 
-
 static unsigned map_hash(const char* str) {
 	unsigned hash = 5381;
 	while (*str) {
-		hash = ((hash << 5) + hash) ^ (unsigned int)(*str++);
+		hash = ((hash << 5) + hash) ^ *str++;
 	}
 	return hash;
 }
@@ -32,20 +31,22 @@ static unsigned map_hash(const char* str) {
 
 static map_node_t* map_newnode(const char* key, void* value, int vsize) {
 	map_node_t* node;
-	int ksize = (int)strlen(key) + 1;
-	int voffset = ksize + (((int)sizeof(void*) - ksize) % (int)sizeof(void*));
-	node = malloc((size_t)((int)sizeof(*node) + voffset + vsize));
+	int ksize = strlen(key) + 1;
+	int voffset = ksize + ((sizeof(void*) - ksize) % sizeof(void*));
+	node = malloc(sizeof(*node) + voffset + vsize);
 	if (!node) return NULL;
-	memcpy(node + 1, key, (size_t)ksize);
+	memcpy(node + 1, key, ksize);
 	node->hash = map_hash(key);
 	node->value = ((char*)(node + 1)) + voffset;
-	memcpy(node->value, value, (size_t)vsize);
+	memcpy(node->value, value, vsize);
 	return node;
 }
 
 
 static int map_bucketidx(map_base_t* m, unsigned hash) {
-	return (int)(hash & (m->nbuckets - 1));
+	/* If the implementation is changed to allow a non-power-of-2 bucket count,
+	 * the line below should be changed to use mod instead of AND */
+	return hash & (m->nbuckets - 1);
 }
 
 
@@ -60,9 +61,9 @@ static int map_resize(map_base_t* m, int nbuckets) {
 	map_node_t* nodes, * node, * next;
 	map_node_t** buckets;
 	int i;
-
+	/* Chain all nodes together */
 	nodes = NULL;
-	i = (int)m->nbuckets;
+	i = m->nbuckets;
 	while (i--) {
 		node = (m->buckets)[i];
 		while (node) {
@@ -72,15 +73,15 @@ static int map_resize(map_base_t* m, int nbuckets) {
 			node = next;
 		}
 	}
-
-	buckets = realloc(m->buckets, (size_t)(sizeof(*m->buckets) * (long unsigned int)nbuckets));
+	/* Reset buckets */
+	buckets = realloc(m->buckets, sizeof(*m->buckets) * nbuckets);
 	if (buckets != NULL) {
 		m->buckets = buckets;
-		m->nbuckets = (unsigned int)nbuckets;
+		m->nbuckets = nbuckets;
 	}
 	if (m->buckets) {
 		memset(m->buckets, 0, sizeof(*m->buckets) * m->nbuckets);
-
+		/* Re-add nodes to buckets */
 		node = nodes;
 		while (node) {
 			next = node->next;
@@ -88,7 +89,7 @@ static int map_resize(map_base_t* m, int nbuckets) {
 			node = next;
 		}
 	}
-
+	/* Return error code if realloc() failed */
 	return (buckets == NULL) ? -1 : 0;
 }
 
@@ -112,7 +113,7 @@ static map_node_t** map_getref(map_base_t* m, const char* key) {
 void map_deinit_(map_base_t* m) {
 	map_node_t* next, * node;
 	int i;
-	i = (int)m->nbuckets;
+	i = m->nbuckets;
 	while (i--) {
 		node = m->buckets[i];
 		while (node) {
@@ -134,17 +135,17 @@ void* map_get_(map_base_t* m, const char* key) {
 int map_set_(map_base_t* m, const char* key, void* value, int vsize) {
 	int n, err;
 	map_node_t** next, * node;
-
+	/* Find & replace existing node */
 	next = map_getref(m, key);
 	if (next) {
-		memcpy((*next)->value, value, (size_t)vsize);
+		memcpy((*next)->value, value, vsize);
 		return 0;
 	}
-
+	/* Add new node */
 	node = map_newnode(key, value, vsize);
 	if (node == NULL) goto fail;
 	if (m->nnodes >= m->nbuckets) {
-		n =(m->nbuckets > 0) ? (int) (m->nbuckets << 1) : 1;
+		n = (m->nbuckets > 0) ? (m->nbuckets << 1) : 1;
 		err = map_resize(m, n);
 		if (err) goto fail;
 	}
@@ -171,7 +172,7 @@ void map_remove_(map_base_t* m, const char* key) {
 
 map_iter_t map_iter_(void) {
 	map_iter_t iter;
-	iter.bucketidx = (unsigned int)-1;
+	iter.bucketidx = -1;
 	iter.node = NULL;
 	return iter;
 }
